+++ builtin/ops/mod.rs
@@ -4,8 +4,10 @@
 use microcad_lang::{eval::BuiltinWorkbenchDefinition, resolve::*};
 
 mod boolean;
+mod extrude;
 mod hull;
-mod transform;
+mod rotate;
+mod translate;
 
 /// Creates the builtin `operation` module
 pub fn ops() -> Symbol {
@@ -14,7 +16,7 @@ pub fn ops() -> Symbol {
         .symbol(boolean::union())
         .symbol(boolean::intersection())
         .symbol(hull::hull())
-        .symbol(transform::Translate::symbol())
-        .symbol(transform::Rotate::symbol())
+        .symbol(translate::Translate::symbol())
+        .symbol(rotate::Rotate::symbol())
         .build()
 }
