+++ builtin/ops/rotate.rs
