--- builtin/ops/transform/mod.rs
@@ -1,10 +1,0 @@
-// Copyright © 2025 The µcad authors <info@ucad.xyz>
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-//! Builtin transformation module.
-
-mod rotate;
-mod translate;
-
-pub use rotate::*;
-pub use translate::*;
